backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
langgraph==0.2.45
langchain-openai==0.2.4
langchain-core==0.3.15
openai==1.54.3
httpx==0.27.2
python-multipart==0.0.12
